requirements.txt
# Hebrew Transcription Pipeline Dependencies

# ============================================
# Core Whisper Dependencies
# ============================================
openai-whisper>=20231117
faster-whisper>=1.0.0

# ============================================
# FastAPI & Web Framework
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# Database
# ============================================
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# ============================================
# Task Queue
# ============================================
celery[redis]>=5.3.0
redis>=5.0.0

# ============================================
# Audio Processing
# ============================================
numpy
torch
tqdm

# Diarization (optional - requires HuggingFace token)
pyannote.audio>=3.1.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
testcontainers>=3.7.0

# ============================================
# Development Tools
# ============================================
ruff>=0.1.0
mypy>=1.8.0
